# swagger/swagger-public.yaml
...
openapi: "3.0.2"
...
info:
  title: "PlacePod API"
  description: "REST services for managing your PlacePod Vehicle Detection Sensor."
  version: "1.0.0"
servers:
 - url: "http://localhost:3000/api/v1"

# Assume global security on all methods unless specifically noted.
security:
  - "bearerToken": []
  

# Method groups
...
paths:
...
  "/lanes":
    "get":
      tags: ["Lane"]
      
      summary: "Returns a list of all lanes."
      description: ""
      operationId: "GetLanes"
      
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/lanes'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
    "post":
      tags: ["Lane"]
      
      summary: "Create a new lane."
      description: ""
      operationId: "CreateLane"
      
      requestBody:
        content:
          "application/json":
            schema:
              $ref: '#/components/schemas/createLane'
      
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/createResponse'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
  "/lanes/{id}":
    "get":
      tags: ["Lane"]
      summary: "Returns the lane matching the given Id."
      description: ""
      operationId: "GetLane"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the lane to get."
        required: true
        schema:
          type: string
          
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/lane'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
          
    "put":
      tags: ["Lane"]
      summary: "Update the lane matching the given Id."
      description: ""
      operationId: "UpdateLane"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the lane to update."
        required: true
        schema:
          type: string
      
      requestBody:
        content:
          "application/json":
            schema:
              $ref: '#/components/schemas/updateLane'
        
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/responseMessage'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
          
    "delete":
      tags: ["Lane"]
      summary: "Delete the lane matching the given Id."
      description: ""
      operationId: "DeleteLane"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the lane to delete."
        required: true
        schema:
          type: string
          
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/responseMessage'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
  "/lanes/{id}/sensors":
    "get":
      tags: ["Lane"]
      summary: "Returns a list of all sensors that belong to the lane matching the given Id."
      description: ""
      operationId: "GetSensorsByLane"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the parent lane."
        required: true
        schema:
          type: string
      
      - $ref: "#/components/parameters/sensorState"
      - $ref: "#/components/parameters/sensorType"
          
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/sensors'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
  "/lanes/{id}/sensorlogs":
      "get":
        tags: ["Lane"]
        summary: "Get PlacePod history by date range for all sensors in a lane matching the given Id."
        description: " - Logs will be returned in descending order.\n - Default filters are a starting time of 4 hours ago, an ending time of now, and a limit of 100 logs.\n - A maximum of 1,000 logs can be returned at once."
        operationId: "GetSensorLogsByLane"
        
        parameters:
        - name: "id"
          in: path
          description: "Id of the lane to get logs for."
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/sensorLogStartTime"
        - $ref: "#/components/parameters/sensorLogEndTime"
        - $ref: "#/components/parameters/sensorLogLimit"
            
        responses:
          "200":
            description: "Success"
            content:
              "application/json":
                schema:
                  $ref: '#/components/schemas/sensorLogs'
          "400":
            $ref: '#/components/schemas/badRequest'
          "401":
            $ref: '#/components/schemas/unauthorized'
          "404":
            $ref: '#/components/schemas/notFound'

# -- End Lane --

# -- Start Oauth2 --
...
  "/parkinglots":

    "get":
      tags: ["Parking Lot"]
      summary: "Returns a list of all parking lots."
      description: ""
      operationId: "GetParkingLots"
      
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/parkingLots'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'

    "post":
      tags: ["Parking Lot"]
      summary: "Create a new parking lot."
      description: ""
      operationId: "CreateLot"

      requestBody:
        content:
          "application/json":
            schema:
              $ref: '#/components/schemas/createLot'
      
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/createResponse'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
  "/parkinglots/{id}":
    "get":
      tags: ["Parking Lot"]
      summary: "Returns the parking lot matching the given Id."
      description: ""
      operationId: "GetParkingLot"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the parking lot to get."
        required: true
        schema:
          type: string
          
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/parkingLot'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
          
    "put":
      tags: ["Parking Lot"]
      summary: "Update the parking lot matching the given Id."
      description: ""
      operationId: "UpdateParkingLot"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the parking lot to update."
        required: true
        schema:
          type: string
      
      requestBody:
        content:
          "application/json":
            schema:
              $ref: '#/components/schemas/updateLot'
        
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/responseMessage'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
          
    "delete":
      tags: ["Parking Lot"]
      summary: "Delete the parking lot matching the given Id."
      description: ""
      operationId: "DeleteParkingLot"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the parking lot to delete."
        required: true
        schema:
          type: string
          
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/responseMessage'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
  "/parkinglots/{id}/lanes":
    "get":
      tags: ["Parking Lot"]
      summary: "Returns a list of all lanes that belong to the parking lot matching the given Id."
      description: ""
      operationId: "GetLanesByParkingLot"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the parent parking lot."
        required: true
        schema:
          type: string

      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/lanes'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
  "/parkinglots/{id}/sensors":
    "get":
      tags: ["Parking Lot"]
      summary: "Returns a list of all sensors that belong to the parking lot matching the given Id."
      description: ""
      operationId: "GetSensorsByParkingLot"
      
      parameters:
      - name: "id"
        in: path
        description: "Id of the parent parking lot."
        required: true
        schema:
          type: string
      
      - $ref: "#/components/parameters/sensorState"
      - $ref: "#/components/parameters/sensorType"
          
      responses:
        "200":
          description: "Success"
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/sensors'
        "400":
          $ref: '#/components/schemas/badRequest'
        "401":
          $ref: '#/components/schemas/unauthorized'
        "404":
          $ref: '#/components/schemas/notFound'
  "/parkinglots/{id}/sensorlogs":
      "get":
        tags: ["Parking Lot"]
        summary: "Get PlacePod history by date range for all sensors in a parking lot matching the given Id."
        description: " - Logs will be returned in descending order.\n - Default filters are a starting time of 4 hours ago, an ending time of now, and a limit of 100 logs.\n - A maximum of 1,000 logs can be returned at once."
        operationId: "GetSensorLogsByParkingLot"
        
        parameters:
        - name: "id"
          in: path
          description: "Id of the parking lot to get logs for."
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/sensorLogStartTime"
        - $ref: "#/components/parameters/sensorLogEndTime"
        - $ref: "#/components/parameters/sensorLogLimit"
            
        responses:
          "200":
            description: "Success"
            content:
              "application/json":
                schema:
                  $ref: '#/components/schemas/sensorLogs'
          "400":
            $ref: '#/components/schemas/badRequest'
          "401":
            $ref: '#/components/schemas/unauthorized'
          "404":
            $ref: '#/components/schemas/notFound'

# -- End ParkingLot


# -- Start Sensors --     
...
components:
...
  schemas:
...
    "createLane":
      title: "Create Lane"
      description: "Fields for creating a new lane."
      type: object
      properties:
        "name":
          type: string
          description: "The name of the lane."
        "frontId":
          type: string
          description: "Id of the sensor in the front position."
          pattern: "^[A-Fa-f0-9]{16}$"
        "backId":
          type: string
          description: "Id of the sensor in the back position."
          pattern: "^[A-Fa-f0-9]{16}$"
        "parkingLotId":
          type: string
          description: "Id of the Parking Lot that this is associated to."
        "direction":
          type: boolean
          description: "Is this lane going into or out of the parking lot?\n\ntrue = in, false = out."
        "count":
          type: number
          format: uint32
          description: "The initial number of vehicles. Defaults to 0 if excluded."
      required: ["name", "frontId", "backId", "parkingLotId", "direction"]
    "createLot":
      title: "Creating Parking Lot"
      description: "Fields for creating a new parking lot."
      type: object
      properties:
        "name":
          type: string
          description: "The name of the parking lot."
        "totalSpaces":
          type: number
          format: uint32
          description: "The number of parking spaces that the parking lot contains."
        "count":
          type: number
          format: uint32
          description: "The estimated number of vehicles in the parking lot at creation. Defaults to 0."
      required: ["name"]
...
    "updateLot":
      title: "Update Parking Lot"
      description: "Fields for updating a parking lot."
      type: object
      properties:
        "name":
          type: string
          description: "The name of the parking lot."
          
        "totalSpaces":
          type: number
          format: uint32
          description: "The number of parking spaces that the parking lot contains."
        "count":
          type: number
          format: uint32
          description: "The estimated number of vehicles in the parking lot."
    "updateLane":
      title: "Update Lane"
      description: "Fields for updating a lane."
      type: object
      properties:
        "name":
          type: string
          description: "The name of the lane."
          
        "frontId":
          type: string
          description: "Id of the sensor in the front position."
          pattern: "^[A-Fa-f0-9]{16}$"

        "backId":
          type: string
          description: "Id of the sensor in the back position."
          pattern: "^[A-Fa-f0-9]{16}$"
          
        "parkingLotId":
          type: string
          description: "Id of the Parking Lot that this is associated to."

        "direction":
          type: boolean
          description: "Is this lane going into or out of the parking lot?\n\ntrue = in, false = out."
          
        "count":
          type: number
          format: uint32
          description: "The initial number of vehicles. Defaults to 0 if excluded."
...
    "responseMessage":
      title: "Response Message"
      type: object
      properties:
        "message":
          type: string
          description: "Response message from the server."
    "createResponse":
      title: "Create Response"
      type: object
      properties:
        "id":
          type: string
          description: The id of the newly created object.
    "lane":
      title: "Lane"
      type: object
      properties:
        "id":
          type: string
          description: "Unique id of the lane."
        "name":
          type: string
          description: "The name of the lane."
        "frontId":
          type: string
          description: "Id of the Sensor in the front position."
          pattern: "^[A-Fa-f0-9]{16}$"
        "backId":
          type: string
          description: "Id of the Sensor in the back position."
          pattern: "^[A-Fa-f0-9]{16}$"
        "parkingLotId":
          type: string
          description: "Id of the Parking Lot that this is associated to."
        "direction":
          type: boolean
          description: "Is this lane going into or out of the parking lot?\n\ntrue = in, false = out."
        "count":
          type: number
          format: uint32
          description: "The last estimated count of vehicles that this Lane has recorded."
    "lanes":
      title: "Lanes"
      type: array
      uniqueItems: true
      items:
        $ref: '#/components/schemas/lane'
    "parkingLot":
      title: "Parking Lot"
      type: object
      properties:
        "id":
          type: string
          description: "Unique id of the parking lot."
        "name":
          type: string
          description: "The name of the parking lot."
        "totalSpaces":
          type: number
          format: uint32
          description: "The number of parking spaces that the parking lot contains."
        "count":
          type: number
          format: uint32
          description: "The estimated number of vehicles in the lot. This is calculated *ONLY* using vehicle counting sensors. This defaults to 0."
    "parkingLots":
      title: "Parking Lots"
      type: array
      uniqueItems: true
      items:
        $ref: '#/components/schemas/parkingLot'
    "sensor":
      title: "Sensor"
      type: object
      properties:
        "id":
          type: string
          description: "Unique Sensor ID. This should match the serial number on the PlacePod unit."
        "name":
          type: string
          description: "Name of the sensor/parking spot."
        "type":
          type: string
          description: "User set sensor type."
        "parkingLotId":
          type: string
          description: "Unique id of the parking lot that the sensor is in."
        "mode":
          type: string
          description: "What type of mode the sensor reports. Currently two possibilities:\n - \"Presence\"\n - \"Counter\""
        "status":
          type: number
          format: uint8
          description: "The current status of the sensor.\n\n If mode is \"Presence\":\n - Status indicates if a car is detected.\n - 0 = No Car Present (Vacant).\n - 1 = Parked Car Detected (Occupied).\n\n If mode is \"Counter\":\n - Status is the last recorded vehicle count.\n - Values range from [0, 128].\n - If the last value was 127, then the next value will roll over to 0.\n - A value of 128 indicates that the PlacePod has been recalibrated or rebooted and has an internal count of 0."
        "temperature":
          type: number
          format: double
          description: "Temperature in Celsius."
        "battery":
          type: number
          format: double
          description: "Battery Voltage."
        "rssi":
          type: number
          format: double
          description: "Sensor’s current RSSI status."
        "snr":
          type: number
          format: double
          description: "Sensor’s current SNR status."
        "serverTime":
          type: string
          format: date-time
          description: "Time the last event was received in UTC.\n - This will represent the last time this sensor saw an uplink message."
        "gatewayTime":
          type: string
          format: date-time
          description: "Time from gateway. This may not be accurate since gateway may have clock skew and/or incorrect time settings."
    "sensors":
      title: "Sensors"
      type: array
      uniqueItems: true
      items:
        $ref: '#/components/schemas/sensor'
    "sensorLog":
      title: "Sensor Log"
      type: object
      properties:
        "sensorId":
          type: string
          pattern: "^[A-Fa-f0-9]{16}$"
          description: "Unique Sensor ID. This should match the serial number on the PlacePod unit."
        "mode":
          type: string
          description: "The type of mode the sensor reported. Currently two possibilities:\n - \"Presence\"\n - \"Counter\""
        "status":
          type: number
          format: uint8
          description: "The status of the sensor.\n\n If mode is \"Presence\":\n - Status indicates if a car is detected.\n - 0 = No Car Present (Vacant).\n - 1 = Parked Car Detected (Occupied).\n\n If mode is \"Counter\":\n - Status is current vehicle count.\n - Values range from [0, 128].\n - If the last value was 127, then the next value will roll over to 0.\n - A value of 128 indicates that the PlacePod has been recalibrated or rebooted and has an internal count of 0."
        "keepAlive":
          type: boolean
          description: "Was the log a keep alive message?"
        "temperature":
          type: number
          format: double
          description: "Temperature in Celsius.\n - This is set up once an hour only if it has changed by +/- 3.0°C, so this may be null."
        "battery":
          type: number
          format: double
          description: "Battery Voltage.\n - This is set up once an hour only if it has changed by +/- 0.01v, so this may be null."
        "frameCount":
          type: number
          format: uint32
          description: "The frame count of this message."
        "rssi":
          type: number
          format: double
          description: "Sensor’s current RSSI status."
        "snr":
          type: number
          format: double
          description: "Sensor’s current SNR status."
        "serverTime":
          type: string
          format: date-time
          description: "Time the event was received in UTC."
        "gatewayTime":
          type: string
          format: date-time
          description: "Time from gateway. This may not be accurate since gateway may have clock skew and/or incorrect time settings."
    "sensorLogs":
        title: "Sensor Logs"
        type: array
        uniqueItems: true
        items:
          $ref: '#/components/schemas/sensorLog'
...
    "badRequest":
      title: "Bad Request"
      description: "Bad Request"
    "unauthorized":
      title: "Unauthorized"
      description: "Unauthorized"
...
    "notFound":
      title: "Not Found"
      description: "Not Found"
  parameters:
    "sensorState":
      name: "state"
      in: query
      description: "Filter for all sensors that are in the given state:\n * **Vacant**: presence sensors that currently do not detect a vehicle.\n * **Occupied**: presence sensors that currently detect a vehicle.\n * **Offline**: any sensors that haven't communicated within the last 24 hours."
      schema:
        type: string
        enum: ["Vacant", "Occupied", "Offline"]
    "sensorType":
      name: "type"
      in: query
      description: "Filter for sensors of the given type."
      schema:
        type: string
    "sensorLogStartTime":
      name: "startTime"
      in: query
      description: "The starting date-time. Defaults to 4 hours ago."
      example: "2018-12-17T00:00:00.000Z"
      schema:
        type: string
        format: date-time
    "sensorLogEndTime":
      name: "endTime"
      in: query
      description: "The end date-time. Defaults to now."
      example: "2018-12-17T04:00:00.000Z"
      schema:
        type: string
        format: date-time
    "sensorLogLimit":
      name: "limit"
      in: query
      description: "Maximum number of logs to return. Defaults to 100. Cap is 1,000 logs."
      example: 100
      schema:
        type: number
        format: uint
      

  # Define what security looks like
  securitySchemes:
     "bearerToken":
      type: http
      description: "Authorization token to send with request."
      scheme: bearer
      bearerFormat: "Bearer 12345"